separate ninja-nitro test setup
