+++ packages/nitro-protocol/package.json
@@ -93,7 +93,7 @@
     "prettier:write": "prettier --write './contracts/**/*.sol'",
     "test": "yarn test:contracts && yarn test:app --all",
     "test:app": "jest -c ./config/jest/jest.config.js",
-    "test:ci": "yarn test:ci:app && yarn test:ci:contracts",
+    "test:ci": "yarn test:ci:app && yarn test:ci:contracts && yarn test:ninja-nitro",
     "test:ci:app": "yarn test:app --all --ci --bail --maxWorkers=4",
     "test:ci:contracts": "yarn test:contracts --all --ci --bail --maxWorkers=4",
     "test:contracts": "jest -c ./config/jest/jest.contracts.config.js",
